Move sonic references out of ep_common
