+++ lib/levanter/src/levanter/grug/_moe/ep_fixed_pooled_wave_all_to_all.py
@@ -17,9 +17,8 @@ from levanter.grug._moe.ep_common import (
     _assignment_sources,
     _ranks_within_groups,
     _token_sources,
-    _use_sonic_gather_sum,
 )
-from levanter.grug._moe.sonic import sonic_gather_sum_masked
+from levanter.grug._moe.sonic import sonic_gather_sum_available, sonic_gather_sum_masked
 from levanter.grug.sharding import _batch_axes
 
 
@@ -54,7 +53,7 @@ def _dispatch_gather_input_grad(
     linear_indices = linear_indices.reshape(tokens_per_shard, topk)
     keep = keep.reshape(tokens_per_shard, topk)
 
-    if _use_sonic_gather_sum():
+    if sonic_gather_sum_available():
         return sonic_gather_sum_masked(cotangent, linear_indices, keep.astype(jnp.float32))
 
     grad_x = jnp.zeros((tokens_per_shard, hidden_dim), dtype=jnp.float32)
@@ -79,7 +78,7 @@ def _combine_gather_sum_impl(
     gather_indices = gather_indices.reshape(tokens_per_shard, topk)
     keep = keep.reshape(tokens_per_shard, topk)
 
-    if _use_sonic_gather_sum():
+    if sonic_gather_sum_available():
         weights = jnp.where(keep, combine_weights, 0)
         return sonic_gather_sum_masked(send_output, gather_indices, weights, output_dtype=jnp.float32)
