+++ lib/levanter/src/levanter/grug/_moe/sonic.py
@@ -155,10 +155,15 @@ def _require_sonic_deps() -> None:
 
 
 def sonic_gather_sum_available() -> bool:
+    """Whether the fused gather-and-weighted-sum kernel can run in this process.
+
+    The kernel is Triton, so a non-GPU backend rules it out however the imports went.
+    """
     return (
         jt is not None
         and _sonic_token_gather_sum_kernel is not None
         and _sonic_token_gather_sum_bwd_kernel is not None
+        and jax.default_backend() == "gpu"
     )
